+++ app/Http/Controllers/Api/TaskController.php
@@ -136,51 +136,5 @@ public function restoreTask(int $id)
     return $this->success($task,'task restore success');
 }
 
-    /**
-     * Generate and save an Excel file for the task with students' notes and grades.
-     * @param mixed $taskId
-     * @return mixed|\Illuminate\Http\JsonResponse
-     */
-    public function generateExcel($taskId)
-    {
-        $filePath = $this->taskService->generateExcel($taskId);
-
-        return response()->json([
-            'message' => 'Excel file has been saved successfully!',
-            'file_path' => $filePath,
-        ]);
-    }
-
-    //.....................................User With Overdue Tasks........................................
-
-    /**
-     * generate an Excel for export Users With Overdue Tasks
-     * @return mixed|\Illuminate\Http\JsonResponse
-     */
-    public function exportUsersWithOverdueTasks()
-    {
-        $this->taskService->exportUsersWithOverdueTasks();
-
-        return $this->success(null,'Excel file has been saved successfully!');
-        
-    }
-
-    //.....................................Course Report Export....................................
-
-    public function exportCourseReport(Course $course)
-    {
-        $this->taskService->exportCourseReport($course);
-
-        return $this->success(null,'Excel file has been saved successfully!');  
-    }
-
-
- //.....................................Education System Export....................................
-
-    public function exportEducationSystem()
-{
-    $this->taskService->exportEducationSystem();
-
-    return $this->success(null,'Excel file has been saved successfully!');  
-}
+   
 }
